politicos

Conjunto de datos de politicos uruguayos que ocuparon cargos politicos entre 1830 y 2020. Incluye datos de:

Presidentes y Vicepresidentes: 1830 - 2020

Legisladores (Diputados y Senadores): 1830 - 2020

Ministros de Estado: 1904 - 2020

Intendentes: 1925 - 2020

Alcaldes y Concejales: 2010 - 2020

Ministros Corte Electoral: 1830 - 2020

Candidato Presidente y Vicepresidente: 1958 - 2020

data(politicos)

Format

data.frame con 10315 filas y 9 columnas.

Details

Las variables que tiene este conjunto de datos son:

politico

Nombre completo del politico en el formato: 'APELLIDO, Nombre'

partido

Partido al que pertenece el politico en el cargo.

fecha_inicio

Fecha de inicio en el que comienza la gestion en el cargo.

fecha_fin

Fecha de fin en la que finaliza la gestion en el cargo.

legislatura

Legislatura en la que ocupa el cargo. Esta variable es para los legisladores unicamente.

cargo

Tipo de cargo.

status

Si es Titular, Suplente, Nominal...

circunscripcion

Distrito al que pertenece el legislador.

sexo

Sexo. Variable numerica: masculino = 1, femenino = 0.

En las variables fecha_inicio y fecha_fin cuando se tiene solo el anio por convencion se le marca la fecha "1 de marzo de ...". La fuente principal de datos de legisladores es el texto: 'Parlamentarios Uruguayos 1830-2005' que fue digitalizado. Se incorporo a la base los datos de candidaturas presidenciales y de precandidaturas presidenciales (elecciones internas) a los efectos de tener informacion mas precisa sobre las carreras politicas. Estos politicos son la unica excepcion en la base que en esas instancias no ocuparon ningun cargo sino que se postularon.

References

AG-CRR (2006). Parlamentarios Uruguayos 1830-2005. Presidencia de la Asamblea General y del Senado y de la Camara de Representantes.

CRR (1971). Tablas Cronologicas 1830-1971. Camara De Representantes, Montevideo, Uruguay.

Musso LA (1971). Anales del Senado del Uruguay. Cronologia sistematizada., volume 1. Republica Oriental del Uruguay, Poder Legislativo, Camara de Senadores.

Oddone JA (1967). Tablas cronologicas: Poder Ejecutivo-Poder Legislativo 1830-1967. Universidad de la Republica, Montevideo, Uruguay.

Cheroni A (1966). Tablas cronologicas del Gobierno Departamenral de Montevideo 1830-1966. Junta Departamental de Montevideo.